backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic[email]==2.10.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
bcrypt==4.2.1
python-multipart==0.0.6
pyjwt==2.11.0
python-dotenv==1.0.1
